inference-server/requirements.txt
# FastAPI server
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# ONNX Runtime
onnxruntime>=1.16.0

# Utilities
scipy>=1.11.0
